# docs/openapi.yaml
openapi: 3.1.0
info:
  title: API AcessaMapa
  version: 1.0.0
  summary: API pública e moderada para informações de acessibilidade urbana.
  description: |
    Contrato da v1. Respostas públicas usam DTOs explícitos e não incluem e-mail,
    credenciais, dados de deficiência ou campos internos de governança. Novas
    contribuições entram em moderação quando `DEMO_WRITE_MODE=moderated`.
servers:
  - url: /api/v1
    description: Mesma origem da aplicação
...
paths:
...
  /locais:
    get:
      tags: [Locais]
      summary: Lista locais aprovados
      operationId: listarLocais
      security: []
      parameters:
        - $ref: "#/components/parameters/Pagina"
        - $ref: "#/components/parameters/Limite"
        - name: busca
          in: query
          description: Busca textual por nome ou endereço.
          schema:
            type: string
            minLength: 2
            maxLength: 100
        - name: categoria
          in: query
          schema:
            $ref: "#/components/schemas/Categoria"
        - name: recurso
          in: query
          schema:
            $ref: "#/components/schemas/ChaveRecurso"
        - name: estadoRecurso
          in: query
          schema:
            $ref: "#/components/schemas/EstadoRecurso"
        - name: oeste
          in: query
          description: Longitude mínima do viewport.
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: sul
          in: query
          description: Latitude mínima do viewport.
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: leste
          in: query
          description: Longitude máxima do viewport.
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: norte
          in: query
          description: Latitude máxima do viewport.
          schema:
            type: number
            minimum: -90
            maximum: 90
      responses:
        "200":
          description: Página de locais aprovados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocaisEnvelope"
        "422":
          $ref: "#/components/responses/ValidationError"
    post:
      tags: [Locais]
      summary: Envia um novo local para moderação
      operationId: criarLocal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalInput"
      responses:
        "202":
          description: Local recebido como pendente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContribuicaoEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WriteDisabled"
        "422":
          $ref: "#/components/responses/ValidationError"
  /locais/{localId}:
    parameters:
      - $ref: "#/components/parameters/LocalId"
    get:
      tags: [Locais]
      summary: Consulta um local aprovado
      operationId: obterLocal
      security: []
      responses:
        "200":
          description: Local aprovado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocalEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Locais]
      summary: Propõe alteração em um local
      description: Alterações materiais em conteúdo aprovado retornam à moderação.
      operationId: editarLocal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalPatchInput"
      responses:
        "202":
          description: Alteração recebida e, quando material, marcada como pendente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContribuicaoEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Locais]
      summary: Arquiva um local
      operationId: arquivarLocal
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Local arquivado logicamente; histórico preservado.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token de curta duração, mantido apenas em memória no cliente.
    refreshCookie:
      type: apiKey
      in: cookie
      name: acessamapa_refresh
      description: Refresh token opaco em cookie HttpOnly, Secure e SameSite=Lax.
...
  parameters:
    Pagina:
      name: pagina
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limite:
      name: limite
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    LocalId:
      name: localId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ObjectId"
...
  responses:
    ValidationError:
      description: Dados de entrada inválidos.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroEnvelope"
    Unauthorized:
      description: Autenticação ausente, inválida ou sessão revogada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroEnvelope"
    Forbidden:
      description: A conta autenticada não tem permissão para a operação.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroEnvelope"
    WriteDisabled:
      description: Escrita desabilitada porque a demonstração está em modo somente leitura.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroEnvelope"
    NotFound:
      description: Recurso não encontrado ou não visível para esta conta.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroEnvelope"
...
  schemas:
    ObjectId:
      type: string
      pattern: "^[a-fA-F0-9]{24}$"
      examples: [66a111111111111111111111]
    RequestId:
      type: string
      minLength: 8
      maxLength: 128
    ErroEnvelope:
      type: object
      additionalProperties: false
      required: [erro]
      properties:
        erro:
          type: object
          additionalProperties: false
          required: [codigo, mensagem, requestId]
          properties:
            codigo:
              type: string
              pattern: "^[A-Z0-9_]+$"
              examples: [VALIDACAO_FALHOU]
            mensagem:
              type: string
              examples: [Revise os campos informados.]
            detalhes:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [campo, mensagem]
                properties:
                  campo:
                    type: string
                  mensagem:
                    type: string
            requestId:
              $ref: "#/components/schemas/RequestId"
...
    UsuarioPublico:
      type: object
      additionalProperties: false
      required: [id, nome]
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        nome:
          type: string
          minLength: 1
          maxLength: 100
        avatar:
          type: [string, "null"]
          format: uri
          maxLength: 2048
          description: URL do avatar ou nulo quando não definida.
...
    Categoria:
      type: string
      enum: [restaurante, hospital, escola, mercado, transporte, lazer, servico_publico, comercio, hotel, outro]
    EstadoRecurso:
      type: string
      enum: [presente, ausente, desconhecido]
    ChaveRecurso:
      type: string
      enum: [rampa, elevador, banheiroAcessivel, pisoTatil, sinalizacaoBraile, estacionamentoAcessivel, portaLarga, libras, audioDescricao, caoPermitido]
    Recursos:
      type: object
      additionalProperties: false
      required: [rampa, elevador, banheiroAcessivel, pisoTatil, sinalizacaoBraile, estacionamentoAcessivel, portaLarga, libras, audioDescricao, caoPermitido]
      properties:
        rampa:
          $ref: "#/components/schemas/EstadoRecurso"
        elevador:
          $ref: "#/components/schemas/EstadoRecurso"
        banheiroAcessivel:
          $ref: "#/components/schemas/EstadoRecurso"
        pisoTatil:
          $ref: "#/components/schemas/EstadoRecurso"
        sinalizacaoBraile:
          $ref: "#/components/schemas/EstadoRecurso"
        estacionamentoAcessivel:
          $ref: "#/components/schemas/EstadoRecurso"
        portaLarga:
          $ref: "#/components/schemas/EstadoRecurso"
        libras:
          $ref: "#/components/schemas/EstadoRecurso"
        audioDescricao:
          $ref: "#/components/schemas/EstadoRecurso"
        caoPermitido:
          $ref: "#/components/schemas/EstadoRecurso"
    GeoJsonPoint:
      type: object
      additionalProperties: false
      required: [type, coordinates]
      properties:
        type:
          type: string
          const: Point
        coordinates:
          type: array
          description: Longitude e latitude, nesta ordem.
          prefixItems:
            - type: number
              minimum: -180
              maximum: 180
            - type: number
              minimum: -90
              maximum: 90
          minItems: 2
          maxItems: 2
    ContagemRecurso:
      type: object
      additionalProperties: false
      required: [informado, confirmacoes, contestacoes, observacoesPresente, observacoesAusente, ultimaVerificacao]
      properties:
        informado:
          $ref: "#/components/schemas/EstadoRecurso"
        confirmacoes:
          type: integer
          minimum: 0
        contestacoes:
          type: integer
          minimum: 0
        observacoesPresente:
          type: integer
          minimum: 0
        observacoesAusente:
          type: integer
          minimum: 0
        ultimaVerificacao:
          type: [string, "null"]
          format: date-time
    ResumoRecursos:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/ContagemRecurso"
    ResumoAvaliacoes:
      type: object
      additionalProperties: false
      required: [media, total, ultimaVerificacao]
      properties:
        media:
          type: [number, "null"]
          minimum: 1
          maximum: 5
        total:
          type: integer
          minimum: 0
        ultimaVerificacao:
          type: [string, "null"]
          format: date-time
    LocalPublico:
      type: object
      additionalProperties: false
      required: [id, nome, descricao, endereco, categoria, localizacao, recursos, resumoAvaliacoes, autor, fotos, criadoEm, atualizadoEm]
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        nome:
          type: string
        descricao:
          type: string
        endereco:
          type: string
        categoria:
          $ref: "#/components/schemas/Categoria"
        localizacao:
          $ref: "#/components/schemas/GeoJsonPoint"
        recursos:
          $ref: "#/components/schemas/Recursos"
        resumoRecursos:
          $ref: "#/components/schemas/ResumoRecursos"
        resumoAvaliacoes:
          $ref: "#/components/schemas/ResumoAvaliacoes"
        autor:
          $ref: "#/components/schemas/UsuarioPublico"
        fotos:
          type: array
          maxItems: 5
          items:
            type: string
            format: uri
        criadoEm:
          type: string
          format: date-time
        atualizadoEm:
          type: string
          format: date-time
    LocalInput:
      type: object
      additionalProperties: false
      required: [nome, descricao, endereco, categoria, localizacao]
      properties:
        nome:
          type: string
          minLength: 2
          maxLength: 200
        descricao:
          type: string
          minLength: 10
          maxLength: 1000
        endereco:
          type: string
          minLength: 5
          maxLength: 300
        categoria:
          $ref: "#/components/schemas/Categoria"
        localizacao:
          $ref: "#/components/schemas/GeoJsonPoint"
        recursos:
          $ref: "#/components/schemas/Recursos"
        fotos:
          type: array
          maxItems: 5
          items:
            type: string
            format: uri
    LocalPatchInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nome:
          type: string
          minLength: 2
          maxLength: 200
        descricao:
          type: string
          minLength: 10
          maxLength: 1000
        endereco:
          type: string
          minLength: 5
          maxLength: 300
        categoria:
          $ref: "#/components/schemas/Categoria"
        localizacao:
          $ref: "#/components/schemas/GeoJsonPoint"
        recursos:
          $ref: "#/components/schemas/Recursos"
        fotos:
          type: array
          maxItems: 5
          items:
            type: string
            format: uri
    LocalEnvelope:
      type: object
      additionalProperties: false
      required: [local]
      properties:
        local:
          $ref: "#/components/schemas/LocalPublico"
    Paginacao:
      type: object
      additionalProperties: false
      required: [pagina, limite, total, totalPaginas, temProximaPagina]
      properties:
        pagina:
          type: integer
          minimum: 1
        limite:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPaginas:
          type: integer
          minimum: 0
        temProximaPagina:
          type: boolean
    LocaisEnvelope:
      type: object
      additionalProperties: false
      required: [locais, paginacao]
      properties:
        locais:
          type: array
          items:
            $ref: "#/components/schemas/LocalPublico"
        paginacao:
          $ref: "#/components/schemas/Paginacao"
    EstadoModeracao:
      type: string
      enum: [pendente, aprovado, rejeitado, arquivado, resolvida, arquivada]
    ContribuicaoEnvelope:
      type: object
      additionalProperties: false
      required: [id, status, mensagem]
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        status:
          $ref: "#/components/schemas/EstadoModeracao"
        mensagem:
          type: string
...